types: further explicit conversions

+++ timer.v
@@ -128,7 +128,7 @@ fn (mut t Timer) draw() {
 	window_height := t.gg.height-20
 	window_x := (t.gg.width - window_width) / 2
 	window_y := (t.gg.height - window_height) / 2
-	t.gg.draw_rect(window_x, window_y, window_width, window_height, gx.white)
+	t.gg.draw_rect(f32(window_x), f32(window_y), f32(window_width), f32(window_height), gx.white)
 	hour_width := window_height / 24 //window_width / 25// 60 / scale  // 60 min
 	scale := 60.0 / f64(hour_width)
 	mut total := 0
@@ -140,7 +140,7 @@ fn (mut t Timer) draw() {
 		x := f64(window_x) + 30.0
 		y := f64(window_y) + f64(task.start) / scale + 10
 		height := f64(task.end - task.start) / scale
-		t.gg.draw_rect(x, y, hour_width,	height		, task.color)
+		t.gg.draw_rect(f32(x), f32(y), f32(hour_width),	f32(height), task.color)
 		t.ft.draw_text(int(x)+hour_width + 10, int(y)+5, task.name + ' ' + task.duration, gx.TextCfg{ color: task.color })
 		if task.productive {
 			total += task.duration_min
@@ -153,14 +153,14 @@ fn (mut t Timer) draw() {
 			t.ft.draw_text(hour_x - 25, hour_y + 10,
 				'${hour:02d}', time_cfg)
 		}
-		t.gg.draw_line(hour_x, hour_y, hour_x + hour_width, hour_y, gx.gray)
+		t.gg.draw_line(f32(hour_x), f32(hour_y), f32(hour_x + hour_width), f32(hour_y), gx.gray)
 	}
 	// Large left vertical line
-	t.gg.draw_line(window_x + 30, window_y + 10, window_x+30, window_y+10+24*
-		hour_width, gx.gray)
+	t.gg.draw_line(f32(window_x + 30), f32(window_y + 10), f32(window_x+30), f32(window_y+10+24*
+		hour_width), gx.gray)
 	// Large right vertical line
-	t.gg.draw_line(window_x + 30 + hour_width, window_y + 10, window_x+30+hour_width, window_y+10+24*
-		hour_width, gx.gray)
+	t.gg.draw_line(f32(window_x + 30 + hour_width), f32(window_y + 10), f32(window_x+30+hour_width), f32(window_y+10+24*
+		hour_width), gx.gray)
 	// Draw the in the top right corner
 	t.ft.draw_text_def(window_x + window_width - 100, 20, t.date.ymmdd())
 	// Draw total time
